+++ AppCenterCrashes/AppCenterCrashes/Internals/Model/MSACWrapperException.m
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#import "MSACException.h"
+#import "MSACWrapperExceptionModel.h"
 #import "MSACWrapperExceptionInternal.h"
 
 @implementation MSACWrapperException