+++ SasquatchMac/SasquatchMacSwift/AppDelegate.swift
@@ -171,30 +171,7 @@ class AppDelegate: NSObject, NSApplicationDelegate, CrashesDelegate, CLLocationM
   }
 
   func attachments(with crashes: Crashes!, for errorReport: ErrorReport!) -> [ErrorAttachmentLog] {
-    var attachments = [ErrorAttachmentLog]()
-
-    // Text attachment.
-    let text = UserDefaults.standard.string(forKey: "textAttachment") ?? ""
-    if !text.isEmpty {
-      let textAttachment = ErrorAttachmentLog.attachment(withText: text, filename: "user.log")!
-      attachments.append(textAttachment)
-    }
-
-    // Binary attachment.
-    let referenceUrl = UserDefaults.standard.url(forKey: "fileAttachment")
-    if referenceUrl != nil {
-      do {
-        let data = try Data(contentsOf: referenceUrl!)
-        let uti = UTTypeCreatePreferredIdentifierForTag(kUTTagClassFilenameExtension, referenceUrl!.pathExtension as NSString, nil)?.takeRetainedValue()
-        let mime = UTTypeCopyPreferredTagWithClass(uti!, kUTTagClassMIMEType)?.takeRetainedValue() as NSString?
-        let binaryAttachment = ErrorAttachmentLog.attachment(withBinary: data, filename: referenceUrl?.lastPathComponent, contentType: mime! as String)!
-        attachments.append(binaryAttachment)
-        print("Add binary attachment with \(data.count) bytes")
-      } catch {
-        print(error)
-      }
-    }
-    return attachments
+    return PrepareErrorAttachments.prepareAttachments()
   }
     
   // CLLocationManager Delegate